+++ changelog.md
@@ -4,6 +4,8 @@
 
 * The internal setting allowing Baserow to run with the user tables in a separate 
   database has been removed entirely to prevent data integrity issues.
+* Fixed bug where the currently selected dropdown item is out of view from the dropdown
+  window when scrolling with the arrow keys.
   
 ## Released (2021-07-16)